+++ src/core/usecases/tests.rs
@@ -254,7 +254,11 @@ impl EntryGateway for MockDb {
     ) -> RepoResult<Vec<Entry>> {
         unimplemented!();
     }
-    fn most_popular_entry_tags(&self, _pagination: &Pagination) -> RepoResult<Vec<TagFrequency>> {
+    fn most_popular_entry_tags(
+        &self,
+        _params: &MostPopularTagsParams,
+        _pagination: &Pagination,
+    ) -> RepoResult<Vec<TagFrequency>> {
         unimplemented!();
     }
     fn count_entries(&self) -> RepoResult<usize> {